requirements.txt
fastapi
uvicorn
openai
pydantic-ai
python-dotenv
